Added on_entry and on_exit states
This patch adds functions which are executed when the statemachine
transitions from a state A to a state B. First on_exit_state_a is called
(which is defined as an empty trait implementation which can be
overridden) followed by a call to on_entry_state_b. The on_entry and
on_exit functions can be implemented in the Statemachine Context trait
implementation or ignored altogether, since the standard implementation
is empty.

Signed-off-by: Martin Broers <martin.broers@bluebirdpower.com>